+++ src/index.js
@@ -42,6 +42,14 @@ const parsers = {
     hasPragma,
     preprocess: text => text.trim(),
   },
+  "apex-anonymous": {
+    astFormat: "apex",
+    parse,
+    locStart,
+    locEnd,
+    hasPragma,
+    preprocess: text => text.trim(),
+  },
 };
 
 const printers = {
@@ -62,10 +70,12 @@ const printers = {
   },
 };
 
+const CATEGORY_APEX = "apex";
+
 const options = {
   apexStandaloneParser: {
     type: "choice",
-    category: "Global",
+    category: CATEGORY_APEX,
     default: "none",
     choices: [
       {
@@ -89,16 +99,17 @@ const options = {
   },
   apexStandalonePort: {
     type: "int",
-    category: "Global",
+    category: CATEGORY_APEX,
     default: 2117,
     description:
       "The standalone server port to connect to. Only applicable if apexStandaloneParser is true. Default to 2117.",
   },
-  apexAnonymous: {
+  apexInsertFinalNewline: {
     type: "boolean",
-    category: "Global",
-    default: false,
-    description: "Treat the code as anonymous Apex",
+    category: CATEGORY_APEX,
+    default: true,
+    description:
+      "Whether to insert one newline as the last thing in the output. Default to true.",
   },
 };