+++ src/parser.js
@@ -32,6 +32,9 @@ function parseTextWithSpawn(text, anonymous) {
   if (executionError) {
     throw executionError;
   }
+  if (executionResult.status !== 0) {
+    throw new Error(executionResult.stdout.toString());
+  }
 
   return executionResult.stdout.toString();
 }
@@ -153,6 +156,17 @@ function handleMethodDeclaration(location, sourceCode, commentNodes, node) {
   return handleNodeEndedWithCharacter(")")(location, sourceCode, commentNodes);
 }
 
+function handleAnnotation(location, sourceCode, commentNodes, node) {
+  // This is an annotation without parameters, so we can use the identity
+  // location
+  if (!node.parameters || node.parameters.length === 0) {
+    return location;
+  }
+  // If not, we need to use the position of the closing parenthesis after the
+  // parameters
+  return handleNodeEndedWithCharacter(")")(location, sourceCode, commentNodes);
+}
+
 // We need to generate the location for a node differently based on the node
 // type. This object holds a String => Function mapping in order to do that.
 const locationGenerationHandler = {};
@@ -208,6 +222,15 @@ locationGenerationHandler[
   apexTypes.SWITCH_STATEMENT
 ] = handleNodeEndedWithCharacter("}");
 locationGenerationHandler[
+  apexTypes.NEW_LIST_LITERAL
+] = handleNodeEndedWithCharacter("}");
+locationGenerationHandler[
+  apexTypes.NEW_SET_LITERAL
+] = handleNodeEndedWithCharacter("}");
+locationGenerationHandler[
+  apexTypes.NEW_MAP_LITERAL
+] = handleNodeEndedWithCharacter("}");
+locationGenerationHandler[
   apexTypes.VARIABLE_DECLARATIONS
 ] = handleNodeEndedWithCharacter(";");
 locationGenerationHandler[
@@ -216,6 +239,7 @@ locationGenerationHandler[
 locationGenerationHandler[
   apexTypes.METHOD_CALL_EXPRESSION
 ] = handleNodeEndedWithCharacter(")");
+locationGenerationHandler[apexTypes.ANNOTATION] = handleAnnotation;
 locationGenerationHandler[
   apexTypes.METHOD_DECLARATION
 ] = handleMethodDeclaration;
@@ -286,7 +310,7 @@ function handleNodeLocation(node, sourceCode, commentNodes) {
     }
   }
   if (currentLocation) {
-    return Object.assign({}, currentLocation);
+    return { ...currentLocation };
   }
   if (node.loc) {
     return {
@@ -331,8 +355,31 @@ function generateExtraMetadata(
   }
   Object.keys(node).forEach(key => {
     if (typeof node[key] === "object") {
-      if (Array.isArray(node) && parseInt(key, 10) === node.length - 1) {
-        node[key].isLastNodeInArray = true; // So that we don't apply trailing empty line after this node
+      if (Array.isArray(node)) {
+        const keyInt = parseInt(key, 10);
+        if (keyInt === node.length - 1) {
+          node[key].isLastNodeInArray = true; // So that we don't apply trailing empty line after this node
+        } else {
+          // Here we flag a node if its next sibling is on the same line.
+          // The reasoning is that for a block of code like this:
+          // ```
+          // Integer a = 1; Integer c = 2; Integer c = 3;
+          //
+          // Integer d = 4;
+          // ```
+          // We don't want a trailing empty line after `Integer a = 1;`
+          // so we need to mark it as a special node.
+          const currentChildNode = node[keyInt];
+          const nextChildNode = node[keyInt + 1];
+          if (
+            nextChildNode &&
+            nextChildNode.loc &&
+            currentChildNode.loc &&
+            nextChildNode.loc.startLine === currentChildNode.loc.endLine
+          ) {
+            node[keyInt].isNextStatementOnSameLine = true;
+          }
+        }
       }
       generateExtraMetadata(
         node[key],
@@ -347,7 +394,8 @@ function generateExtraMetadata(
     apexClass &&
     nodeLoc &&
     allowTrailingEmptyLine &&
-    !node.isLastNodeInArray
+    !node.isLastNodeInArray &&
+    !node.isNextStatementOnSameLine
   ) {
     const nextLine = nodeLoc.endLine + 1;
     const nextEmptyLine = emptyLineLocations.indexOf(nextLine);
@@ -435,10 +483,13 @@ function parse(sourceCode, _, options) {
       sourceCode,
       options.apexStandaloneHost,
       options.apexStandalonePort,
-      options.apexAnonymous,
+      options.parser === "apex-anonymous",
     );
   } else {
-    serializedAst = parseTextWithSpawn(sourceCode, options.apexAnonymous);
+    serializedAst = parseTextWithSpawn(
+      sourceCode,
+      options.parser === "apex-anonymous",
+    );
   }
   let ast = {};
   if (serializedAst) {