+++ CHANGELOG.md
@@ -1,10 +1,53 @@
-## Unreleased
-- Update dependency requirement for Prettier.
+# Unreleased
+- Fix last comment in Annotations being put in wrong location ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/142)).
 
-## 1.0.0-rc.2
+# 1.0.0
+
+## API Changes
+- Remove `--apex-anonymous` option, use `apex-anonymous` parser instead ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/106)).
+
+## Formatting Changes
+- By default, add curly bracket spacing for maps, lists and sets ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/99)).
+- Fix extraneous linebreaks being added after `this` variable expression ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/98)).
+- Fix wrong comment position in a long method/variable chain ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/116)).
+- Fix extraneous linebreaks for empty parameter list in long method declarations ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/124)).
+- Fix SOQL query numbers losing negative value ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/122)).
+- Fix method call expressions containing `this` and `super` having an extra indentation level.
+- Fix trailing comment wrong location after last element in list/set/map init literals ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/126)).
+
+# 1.0.0-rc.6
+
+- Fix incorrect format for decimals/doubles ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/88)).
+- Add support for Number Expression in SOQL Geolocation Literal ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/90)).
+- Fix duplicated trailing empty lines for multiple expressions on the same line ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/91)).
+- Fix expressions in Name Value Parameters not being indented ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/93)).
+- Fix SOQL queries in binaryish expressions having extraneous indentations ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/96)).
+- Fix SOQL query number being printed as different types (`big-decimal` to `int`) in certain situations.
+
+# 1.0.0-rc.5
+
+- CLI/Option change:
+  - Add `apex-insert-final-newline` option ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/83)).
+- Remove newline before SOQL expression in ForInit ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/28)).
+- Add space before colon in ForEnhancedControl ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/86)).
+
+# 1.0.0-rc.4
+
+- Add support for AnnotationString ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/80)).
+
+# 1.0.0-rc.3
+
+- Update dependency requirement for Prettier.
+- Fix incorrect format for SOQL query with multiple types in FROM clause ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/76)).
+- Fix incorrect order of expressions in SOSL query ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/75)).
+- Add support for class and interface generics ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/77)).
+
+# 1.0.0-rc.2
+
 - Remove references to deleted vendor file `apex-ast-serializer-ng`.
 
-## 1.0.0-rc.1
+# 1.0.0-rc.1
+
 - CLI/Option change:
   - Remove `apex-verify-ast` option. Please use `--debug-check` instead.
 - Implement comments using Prettier's API.
@@ -20,18 +63,20 @@
 - Fix unstable leading comments to ValueWhen and ElseWhen blocks.
 - Throw errors when encountering unknown node types.
 
-## 1.0.0-beta.2
+# 1.0.0-beta.2
+
 - Fix long static method calls producing undeployable code ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/69)).
 
-## 1.0.0-beta.1
+# 1.0.0-beta.1
+
 - Add support for Anonymous Code block with `--apex-anonymous` option.
 - CLI/Option change:
   - `use-standalone-server` option is now `apex-standalone-parser`,
-  and it is now a choice between `none` and `built-in`.
+    and it is now a choice between `none` and `built-in`.
   - `server-port` option is now `apex-standalone-port`.
   - Add `apex-verify-ast` option.
 - Fix dangling comments being printed incorrectly for Triggers ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/13)).
-Thanks to @praksb, @ntotten and @vazexqi for their help on getting jorje fixed.
+  Thanks to @praksb, @ntotten and @vazexqi for their help on getting jorje fixed.
 - Fix SOQL unary expression not generating space before next expression.
 - Add support for SOQL WHERE Calculation Expression ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/34)).
 - Add support for parameter modifiers ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/37)).
@@ -56,7 +101,8 @@ Thanks to @praksb, @ntotten and @vazexqi for their help on getting jorje fixed.
 - Fix unstable NameValueParameter trailing comments.
 - Fix unstable WhereOpExpr trailing comments.
 
-## 1.0.0-alpha.9
+# 1.0.0-alpha.9
+
 - Fix trailing comments after class names not being printed ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/20)).
 - Add new lines in empty blocks for Enum.
 - Add new lines to long list of annotation parameters.
@@ -68,44 +114,47 @@ Thanks to @praksb, @ntotten and @vazexqi for their help on getting jorje fixed.
 - Fix inline comments in blocks being attached to wrong node ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/23)).
 - Fix comments being pushed to next lines because of node's trailing empty line ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/26)).
 
-## 1.0.0-alpha.8
+# 1.0.0-alpha.8
+
 - Fix comments not being printed if they appear before the root node ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/17)).
 - Fix dangling comments not being printed for Class, Interface, Enum and Block Statement ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/13)).
 - Add new lines in empty blocks ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/15)).
 - Fix trailing comment being mistaken as leading comment in certain situations,
-because `jorje` provides wrong start and end indexes for some node types ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/19)).
+  because `jorje` provides wrong start and end indexes for some node types ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/19)).
 
-## 1.0.0-alpha.7
+# 1.0.0-alpha.7
+
 - Fix Annotated Declaration being indented too far ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/18)).
 - Getter and setter will no longer break if they can fit on the same line.
 - Add documentation for configuration options.
 
-## 1.0.0-alpha.6
+# 1.0.0-alpha.6
+
 - Fix formatting for trailing comments at the end of block statements ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/12)).
 - Fix issue in which comments right before semi colon are not printed ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/11)).
 - Assert that all comments have been printed out in the formatted code.
 - Fix long expressions not breaking into multi lines ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/16)).
 
-## 1.0.0-alpha.5
+# 1.0.0-alpha.5
 
-- Fix `apex-ast-serializer` executables not having their execute bits set on *nix ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/10)).
+- Fix `apex-ast-serializer` executables not having their execute bits set on \*nix ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/10)).
 
-## 1.0.0-alpha.4
+# 1.0.0-alpha.4
 
 - Support `WITH SECURITY_ENFORCED` in SOQL ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/9)).
 - Fix `npm scripts` pointing to old files.
 
-## 1.0.0-alpha.3
+# 1.0.0-alpha.3
 
 - Fix DML operation having double indents ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/8)).
 - Rename scripts to start and stop Apex Parser Server.
 
-## 1.0.0-alpha.2
+# 1.0.0-alpha.2
 
 - Use Prettier's default options for `tabWidth` and `printWidth`.
 - Invoke `apex-ast-serializer` directly by default,
-with option to use Nailgun server.
+  with option to use Nailgun server.
 
-## 1.0.0-alpha.1
+# 1.0.0-alpha.1
 
 - Initial release.